+++ lib/evil_events/core/events/abstract_event.rb
@@ -25,8 +25,8 @@ module EvilEvents::Core::Events
     include EventExtensions::ClassSignature
     # @since 0.3.0
     include EventExtensions::Hookable
-    # @since 0.1.0
-    extend EvilEvents::Shared::CombinedContext::Mixin
+    # @since 0.4.0
+    extend Symbiont::Context
 
     # @return [String]
     #